Add a room object.
Each creature has a weak pointer to a room.  Rooms will have weak
references to each other.  Rooms keep a strong shared pointer to
creatures in them.  The driver will keep strong shared pointers to
rooms, cleaning as desired.  When a room gets destructed, creatures
in it that are not pointed to by anything else will get destructed
as well.